+++ R/das_process.R
@@ -13,7 +13,8 @@
 #'   and thus also when state/condition information
 #'   (cruise number, weather, Bft, Mode, etc) is reset
 #' @param reset.event logical; default is \code{TRUE}.
-#'   Indicates if state/condition information (weather, Bft, Mode, etc) should be reset to \code{NA}
+#'   Indicates if state/condition information (weather, Bft, Mode, etc)
+#'   should be reset to \code{NA}
 #'   if there is an applicable event with an \code{NA} for that state/condition
 #' @param reset.effort logical; default is \code{TRUE}.
 #'   Indicates if state/condition information should be reset to \code{NA}
@@ -22,7 +23,8 @@
 #'   Indicates if state/condition information should be reset to \code{NA}
 #'   at the beginning of each day. This argument should only
 #'   be set to \code{FALSE} for comparison with older methods, such as REPORT
-#' @param add.dtll.sight logical indicating if the DateTime (dt) and latitude and longitude (ll)
+#' @param add.dtll.sight logical indicating if the DateTime (dt)
+#' and latitude and longitude (ll)
 #'   columns should be added to the sighting events (?, 1, 2, 3, 4, 5, 6, 7, and 8)
 #'   from the corresponding (immediately preceding) A event
 #'
@@ -30,7 +32,8 @@
 #'   it is assumed to be a filepath and first passed to \code{\link{das_read}}.
 #'   This output is then passed to \code{das_process}.
 #'
-#'   DAS data is event-based, meaning most events indicate when a state or weather condition changes.
+#'   DAS data is event-based, meaning most events indicate when a
+#'   state or weather condition changes.
 #'   For instance, a 'V' event indicates when one or more sea state viewing conditions
 #'   (such as Beaufort sea state) change, and these conditions
 #'   are the same for subsequent events until the next 'V' event.
@@ -49,8 +52,9 @@
 #'     \item r events are converted to R events with non-standard effort;
 #'       see \code{\link{das_format_pdf}} for more details
 #'     \item An event is considered 'on effort' if it is 1) an R event,
-#'       2) a B event immediately preceding an R event, or 3) between corresponding R and E events
-#'       (not including the E event). The 'EffortDot' column is not used when determining on effort data.
+#'       2) a B event immediately preceding an R event, or
+#'       3) between corresponding R and E events (not including the E event).
+#'       The 'EffortDot' column is not used when determining on effort data.
 #'       Note that effort is reset to 'off effort' at the beginning of a new day.
 #'     \item All state/condition information is reset at the beginning of each cruise.
 #'       New cruises are identified using \code{days.gap}.
@@ -60,6 +64,14 @@
 #'       An event sequence means that all of the events have the same Lat/Lon/DateTime info,
 #'       and thus previous values for conditions set during the event sequence should not
 #'       carry over to any part of the event sequence.
+#'     \item 'OffsetGMT' is converted to an integer. Values are expected to be
+#'       consistent within a day for each cruise, so events will have an
+#'       OffsetGMT value if there is any B event with the offset data on
+#'       the same day, whether that event is before or after the B event.
+#'       Thus, if any date/cruise combinations have multiple OffsetGMT values
+#'       in the data, then a warning message will be printed and the
+#'       OffsetGMT values will be all NA (for the entire output).
+#'
 #'     \item 'Mode' is capitalized, and 'Mode' values of \code{NA} are assigned a value of "C"
 #'     \item 'EffType' is capitalized, and values of \code{NA} are assigned a value of "S"
 #'     \item 'ESWsides' represents the number of sides being searched during that effort section -
@@ -81,6 +93,7 @@
 #'     On/off effort                 \tab OnEffort  \tab B/R and E events\cr
 #'     Cruise number                 \tab Cruise    \tab Event: B; Column: Data1\cr
 #'     Effort mode                   \tab Mode      \tab Event: B; Column: Data2\cr
+#'     GMT offset of DateTime data   \tab OffsetGMT \tab Event: B; Column: Data3\cr
 #'     Effort type                   \tab EffType   \tab Event: R; Column: Data1\cr
 #'     Number of sides with observer \tab ESWSide   \tab Event: R; Column: Data2\cr
 #'     Course (ship direction)       \tab Course    \tab Event: N; Column: Data1\cr
@@ -98,6 +111,9 @@
 #'     Right observer                \tab ObsR      \tab Event: P; Column: Data3\cr
 #'     Independent observer          \tab ObsInd    \tab Event: P; Column: Data4\cr
 #'   }
+#'
+#'   OffsetGMT represents the difference in hours between the DateTime data
+#'   (which should be in local time) and GMT (i.e., UTC).
 #'
 #'   Internal warning messages are printed with row numbers of the input file
 #'   (NOT of the output data frame) of unexpected event codes and r events,
@@ -231,6 +247,34 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
 
 
   #----------------------------------------------------------------------------
+  ### Extract GMT Offsets for dates + cruise numbers
+  x.offset <- x %>%
+    filter(.data$Event == "B") %>%
+    mutate(Date = as.Date(.data$DateTime),
+           Cruise = as.numeric(.data$Data1),
+           OffsetGMT = as.integer(.data$Data3)) %>%
+    select(.data$Date, .data$Cruise, .data$OffsetGMT) %>%
+    distinct()
+
+  x.offset.summ <- x.offset %>%
+    group_by(.data$Date, .data$Cruise) %>%
+    summarise(offsetgmt_uq = n_distinct(.data$OffsetGMT))
+
+
+
+  if (any(x.offset.summ$offsetgmt_uq != 1)) {
+    x.offset.mult <- x.offset.summ %>% filter(.data$offsetgmt_uq > 1)
+    warning("The following dates + cruise numbers have multiple OffsetGMT ",
+            "values (Field 3 of B events), and thus the OffsetGMT column ",
+            "will contain only NAs in the output data frame:\n",
+            paste(paste(x.offset.mult$Date, x.offset.mult$Cruise, sep = " + "),
+                  collapse = "\n"),
+            immediate. = TRUE)
+    x.offset$OffsetGMT <- NA_integer_
+  }
+
+
+  #----------------------------------------------------------------------------
   # Add columns for helpful info that is pertinent for a series of records,
   #   such as Beaufort but not sigting cue
 
@@ -336,8 +380,8 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
   #--------------------------------------------------------
   ### Post-processing
   tmp <- list(
-    Cruise = Cruise, Mode = Mode, EffType = EffType, ESWsides = ESWsides,
-    Course = Course, SpdKt = SpdKt,
+    Cruise = Cruise, Mode = Mode,
+    EffType = EffType, ESWsides = ESWsides, Course = Course, SpdKt = SpdKt,
     Bft = Bft, SwellHght = SwellHght, WindSpdKt = WindSpdKt,
     RainFog = RainFog, HorizSun = HorizSun, VertSun = VertSun, Vis = Vis,
     OnEffort = Eff,
@@ -394,20 +438,23 @@ das_process.das_dfr <- function(x, days.gap = 20, reset.event = TRUE,
     rm(x.key, x.tmp)
   }
 
+
   #----------------------------------------------------------------------------
   ### Create and order data frame to return
   cols.tokeep <- c(
     "Event", "DateTime", "Lat", "Lon", "OnEffort",
-    "Cruise", "Mode", "EffType", "ESWsides", "Course", "SpdKt",
+    "Cruise", "Mode", "OffsetGMT", "EffType", "ESWsides", "Course", "SpdKt",
     "Bft", "SwellHght", "WindSpdKt",
     "RainFog", "HorizSun", "VertSun", "Glare", "Vis",
     "ObsL", "Rec", "ObsR", "ObsInd",
     paste0("Data", 1:12), "EffortDot", "EventNum", "file_das", "line_num"
   )
 
-  as_das_df(
-    select(data.frame(x, tmp, stringsAsFactors = FALSE), !!cols.tokeep)
-  )
+  data.frame(x, tmp, stringsAsFactors = FALSE) %>%
+    mutate(Date = as.Date(.data$DateTime)) %>%
+    left_join(x.offset, by = c("Cruise", "Date")) %>%
+    select(!!cols.tokeep) %>%
+    as_das_df()
 }