+++ R/swfscDAS-package.R
@@ -11,9 +11,9 @@
 #' @author Sam Woodman \email{sam.woodman@@noaa.gov}
 #' @seealso \url{https://smwoodman.github.io/swfscDAS/}
 #'
-#' @importFrom dplyr add_row arrange between bind_cols bind_rows case_when desc everything
-#'   filter full_join group_by lag left_join mutate n right_join
-#'   select slice starts_with summarise ungroup
+#' @importFrom dplyr add_row arrange between bind_cols bind_rows case_when desc
+#'   distinct everything filter full_join group_by lag left_join mutate n n_distinct
+#'   right_join select slice starts_with summarise ungroup
 #' @importFrom lubridate year month day tz
 #' @importFrom magrittr %>%
 #' @importFrom parallel clusterExport detectCores parLapplyLB stopCluster