+++ tests/testthat/test-output.R
@@ -4,7 +4,7 @@ y.proc <- das_process(y.read)
 
 exp.proc.name <- c(
   "Event", "DateTime", "Lat", "Lon", "OnEffort",
-  "Cruise", "Mode", "EffType", "ESWsides", "Course", "SpdKt",
+  "Cruise", "Mode", "OffsetGMT", "EffType", "ESWsides", "Course", "SpdKt",
   "Bft", "SwellHght", "WindSpdKt",
   "RainFog", "HorizSun", "VertSun", "Glare", "Vis",
   "ObsL", "Rec", "ObsR", "ObsInd",
@@ -86,6 +86,7 @@ test_that("das_process output has expected column names and classes", {
     OnEffort = "logical",
     Cruise = "numeric",
     Mode = "character",
+    OffsetGMT = "integer",
     EffType = "character",
     ESWsides = "numeric",
     Course = "numeric",